+++ BasicPackage/ManufacturingDomain/Foods/FoodMaterial.cs
@@ -14,10 +14,10 @@ namespace InfinityWorldChess.ManufacturingDomain.Foods
 {
     public class FoodMaterial : Item,IAttachProperty,IOverloadedItem
     {
-        [field:S(6)]public byte Living { get; set; }
-        [field:S(6)]public byte Kiling { get; set; }
-        [field:S(6)]public byte Nimble { get; set; }
-        [field:S(6)]public byte Defend { get; set; }
+        [field:S(6)]public int Living { get; set; }
+        [field:S(6)]public int Kiling { get; set; }
+        [field:S(6)]public int Nimble { get; set; }
+        [field:S(6)]public int Defend { get; set; }
         [field:S(7)]public byte Length { get; set; }
         [field:S(32)]public IActionable<CustomFood> StartEffects { get; set; }
         [field:S(32)]public IActionable<CustomFood> FinishEffects { get; set; }
