+++ MainPackage/BuffDomain/IAttachProperty.cs
@@ -3,12 +3,12 @@ namespace InfinityWorldChess.BuffDomain
     public interface IAttachProperty
     {
         // 生杀灵御
-        public byte Living { get; set; }
+        public int Living { get; set; }
 
-        public byte Kiling { get; set; }
+        public int Kiling { get; set; }
 
-        public byte Nimble { get; set; }
+        public int Nimble { get; set; }
 
-        public byte Defend { get; set; }
+        public int Defend { get; set; }
     }
 }
